Lab 4 - Functional Programming
TODO#1 – In the Subscriber class, we need to implement a method saveMedia() that will save all the media object into a serialized file (given a storage path).
TODO#2 – Of course, we then need to implement another method to deserialize the Media object that we’ve previously saved. You can do with the getMedia() method.
TODO#3 – In order to ensure, that the Media objects are properly save, we need to implement Serializable